[Misc] Cleanup more configs and processors
