+++ vllm/model_executor/models/flex_olmo.py
@@ -24,7 +24,7 @@ from vllm.model_executor.layers.fused_moe import FusedMoE
 from vllm.model_executor.layers.layernorm import RMSNorm
 from vllm.model_executor.layers.linear import ReplicatedLinear
 from vllm.model_executor.models.olmoe import OlmoeAttention, OlmoeForCausalLM
-from vllm.transformers_utils.configs import FlexOlmoConfig
+from vllm.transformers_utils.configs.flex_olmo import FlexOlmoConfig
 
 logger = init_logger(__name__)
