+++ vllm/transformers_utils/configs/speculators/__init__.py
@@ -1,2 +1,5 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+from .base import SpeculatorsConfig
+
+__all__ = ["SpeculatorsConfig"]
